pyproject.toml
[tool.poetry]
name = "edu-pl"
version = "0.1.0"
description = ""
authors = ["maetin <maetin0324@gmail.com>"]
readme = "README.md"
packages = [{include = "edu_pl"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.96.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
openai = "^0.27.7"
sqlalchemy = "^2.0.15"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
